Added save/load, added time conversion to ms,
Attempted to add a save and load class to save user input in an EditText that would manipulate the timer. Also added an int to millisecond converter and time converter to get and convert the time that the user inputs into the EditText to change the dryer and washer times.